refactor: combine live & oneOnOne routes and controllers
